+++ pkg/dinkurafk/hooks_windows.c
@@ -17,7 +17,7 @@
 // You should have received a copy of the GNU General Public License along with
 // this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "windows_hooks.h"
+#include "hooks_windows.h"
 #include "_cgo_export.h"
 
 // Global vars
