+++ pkg/dinkurafk/hooks_windows.h
